+++ app/renderer/reducers/Inspector.js
@@ -12,7 +12,7 @@ import { SET_SOURCE_AND_SCREENSHOT, QUIT_SESSION_REQUESTED, QUIT_SESSION_DONE,
          SET_SWIPE_START, SET_SWIPE_END, CLEAR_SWIPE_ACTION, SET_SEARCHED_FOR_ELEMENT_BOUNDS, CLEAR_SEARCHED_FOR_ELEMENT_BOUNDS,
          PROMPT_KEEP_ALIVE, HIDE_PROMPT_KEEP_ALIVE,
          SELECT_ACTION_GROUP, SELECT_SUB_ACTION_GROUP,
-         SELECT_INTERACTION_MODE, ENTERING_ACTION_ARGS, SET_ACTION_ARG, REMOVE_ACTION
+         SELECT_INTERACTION_MODE, ENTERING_ACTION_ARGS, SET_ACTION_ARG, REMOVE_ACTION, SHOW_CHECK_MODAL, HIDE_CHECK_MODAL
 } from '../actions/Inspector';
 import { SCREENSHOT_INTERACTION_MODE, INTERACTION_MODE } from '../components/Inspector/shared';
 
@@ -162,6 +162,22 @@ export default function inspector (state = INITIAL_STATE, action) {
         action: {
           ...state.action,
           sendKeys: null,
+        }
+      };
+
+    case SHOW_CHECK_MODAL:
+      return {
+        ...state,
+        checkModalVisible: true
+      };
+
+    case HIDE_CHECK_MODAL:
+      return {
+        ...state,
+        checkModalVisible: false,
+        action: {
+          ...state.action,
+          check: null,
         }
       };